refactor: extract TrainingDataset to dataset.py; repoint dataset patch targets (#74)
